refactor: Seperates code into modules
